split NaiveQuorumChecker out of QuorumChecker.hpp
